changes in structure os folders
