+++ televendas/apps.py
