+++ televendas/serializers/sale.py
@@ -1,0 +1,7 @@
+from .models import Sale
+from rest_framework import serializers
+
+class SaleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Sale
+        fields = '__all__'
