+++ televendas/serializers/seller.py
@@ -1,0 +1,8 @@
+from .models import Seller 
+from rest_framework import serializers
+
+class SellerSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Seller
+        fields = '__all__'
+
