+++ televendas/urls.py
@@ -1,6 +1,6 @@
 from django.urls import include, path
 from rest_framework import routers
-from .views import ComissionPlanViewSet
+from .views.comission_plan import ComissionPlanViewSet
 
 router = routers.DefaultRouter()
 router.register(r'comissions', ComissionPlanViewSet)
