+++ televendas/views/comission_plan.py
@@ -1,8 +1,7 @@
 from rest_framework import status, viewsets
-from televendas.serializers import ComissionPlanSerializer
-from .models import ComissionPlan
+from televendas.serializers.comission_plan import ComissionPlanSerializer
+from televendas.models.comission_plan import ComissionPlan
 
 class ComissionPlanViewSet(viewsets.ModelViewSet):
     queryset = ComissionPlan.objects.all()
     serializer_class = ComissionPlanSerializer
-
